Create method to write lines of data
I`ve decided to store the data on a CSV file as it is portable and
generally lightweight compared to other formats. This method take a line
od data and writes it to a CSV file.

Moreover, another method to sanke case the platform name was created to
standardize the platform names on the file.