Create a platform factory
That was an elegant solution I have came up with to initialize a
platform instance. This is meant to avoid excessive usage of
if-elif-else blocks and switch statements.

The method get for the globals dictionary could have been used as well,
but I thought that raising an exception while catching another would be
more meaningful