Improve request response for ghost platforms
The term "ghost platform" was designed to refer to the platform that
does not exist in the system. "Ghost" is shorter than "non-existent",
and the last one might not fit on my self-imposed limit of 55 characters
in the commit subject limit.

Keeping the commit subject restricted is considered a good practice due
to the fact that longer commit titles may overflow from some screens,
and reduce visibility.

About the commit itself, it seeks to improve the response message from
"There is no such platform as SitePlatform" to "There is no implemented
platform for Site (SitePlatform)".

Given this software is mean for internal usage, in the next commit I'll
include a tip that may help solving this problem