Create base implementations for platform methods
These base implementations will become handy as I seek to build 2
inherited classes for beautiful soup and selenium backed strategies. Not
every platform will require headers or cookies, and in these cases, I'd
have to use this basic implementation everytime it happened: this is
not a good idea.

Instead, I'm having this basic implementation that will be overridden
only when necessary. This will make the classes less verbose

Moreover, I plan to have a superclass for platforms using, BSoup-4 and
any other future strategy, so if price and title getters were abstract,
I might have to implement them on the strategy class, which is not meant
to happen, because this logic is strictly bound to the platform

For logging reason, I'm implementing an init method announcing the
platform has been initialized