Extract validation method
Previously, there was a method checking for the existence of error key
on the current chunk of data, and this method was inside`filter_errors`
function. In order to improve the system "testability", I've extracted
this method.

If it's just an extraction, why is the new method different from the
previous one?

Well, it deos happen, because, there is a new kind of instance to ignore
 when analyzing the new set of data, and this is the ones without a link
  key in the object