Modify setup validation class name
As `Setup` did not say much about what was done in that class. I really
believe `SetupVerification` is much more meaningful, as it verifies the
json entries being passed by the API and selects the valid lines.

The probabilistic selection is the only exquisite responsibility in this
class, but I believe it might be worse if I moved this method somewhere
else, as it;s part of the setup process