Create README.md file
This is made to help other developers to contribute with this project.

It has instructions to run the application and test it locally with cURL
that has its exit prettified with `json_pp` tool. It's present on
Unix-based OS's by default, but can be installed on Windows. The other
option is to use HTTP clients like Postman, Insomnia or API Dog.

IT has instructions for requests and unit tests being run with pytest,
which has a better CLI than unittest module. So far, it's being better
to write tests with unittest because of the OOP approach, and running it
 with pytest

 PS: pytest and unittest are not in the requirements.txt file, as they
 are "side requirements", in other words, they are not necessary for the
  application to run, but are vital for the tests