Remove lines parameter from API call
In order to make the API call a little easier, I'm removing the line
parameter.

Instead of taking the user input for that, I'm setting default values
for instance selection based on the based platform.

For those using Beautiful Soup, there will be 50 tested products. For
the others using Selenium, 20 products are verified, as it takes more
time for it to run

These values can be modified on a particular platform, if necessary